+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## Unreleased
+
+## Removed
+
+* option for different config has no real use and unnecessarily complicates things
+
+## Fixed
+
+* wrong use of reversed in select-previous,select-previous
+
 ## 0.8.0 (Naming things is hard) - 2017-06-20
 ## Changed